apps/worker/pyproject.toml
[project]
name = "vouchr-worker"
version = "0.1.0"
description = "FastAPI worker for statement extraction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "pydantic-settings>=2.10.1",
  "argon2-cffi>=25.1.0",
  "google-genai>=1.31.0",
  "boto3>=1.40.23",
  "httpx>=0.28.1",
  "pypdf>=6.0.0",
  "pdfplumber>=0.11.0",
  "openai>=1.0.0",
  "PyJWT>=2.0.0",
  "python-multipart>=0.0.20"
]

[tool.uv]
package = false
